+++ MakefileGame
@@ -67,12 +67,12 @@ Lamecraft/Mobs/Models/CowModel.o \
 Lamecraft/Mobs/Models/CreeperModel.o \
 Lamecraft/Mobs/Models/HumanoidModel.o \
 Lamecraft/Mobs/Models/SheepModel.o \
-LameCraft/States/StateMenu.o \
-LameCraft/States/StateOptions.o \
+LameCraft/States/menu/StateMenu.o \
+LameCraft/States/menu/StateOptions.o \
 LameCraft/States/StatePlayCreative.o \
 LameCraft/States/StatePlay.o \
-LameCraft/States/StateLanguageSelect.o \
-LameCraft/States/StateMainMenu.o \
+LameCraft/States/menu/StateLanguageSelect.o \
+LameCraft/States/menu/StateMainMenu.o \
 LameCraft/Screens/LoadingScreen.o \
 LameCraft/Screens/SplashScreen.o \
 LameCraft/Lang/Translation.o \
